librispeech_loader: refactoring
