+++ librispeech/data_loader.py
@@ -15,7 +15,7 @@ class LibriSpeechH5pyTestDataLoader(DataLoader):
 
 if __name__ == "__main__":
     from mics.transforms import get_test_transform
-    from librispeech.h5py_reader import LibriSpeechH5py
+    from librispeech.h5py_utils.h5py_reader import LibriSpeechH5py
 
     test_transforms = get_test_transform(length=2 ** 14)
     test_dataset = LibriSpeechH5py("./librispeach/train-clean-100.hdf5",
