+++ librispeech/h5py_utils/h5py_reader.py
@@ -4,25 +4,12 @@ import numpy as np
 
 from torch.utils import data
 
-
-from mics.utils import tensor_to_numpy
-from librispeech.utils import SPEAKER, SOUND, CHAPTER, UTTERANCE
-from sklearn.preprocessing import LabelEncoder, OneHotEncoder
-
-
-class LabelsToOneHot:
-    def __init__(self, data):
-        self.labels_encoder = LabelEncoder()
-        self.one_hot_encoder = OneHotEncoder()
-
-        self.labels_encoder.fit(data.reshape(-1, ))
-        self.one_hot_encoder.fit(self.labels_encoder.transform(data.reshape(-1, )).reshape((-1, 1)))
-
-    def __call__(self, data):
-        return self.one_hot_encoder.transform(self.labels_encoder.transform(data.reshape(-1, )).reshape((-1, 1)))
+from librispeech.basic_reader import LibriSpeechBasic
+from mics.utils import LabelsToOneHot
+from librispeech.h5py_utils.utils import SPEAKER, SOUND, CHAPTER, UTTERANCE
 
 
-class LibriSpeechH5py(data.Dataset):
+class LibriSpeechH5py(LibriSpeechBasic):
     def __init__(self,
                  dataset_path,
                  transforms,
@@ -34,11 +21,7 @@ class LibriSpeechH5py(data.Dataset):
                  one_hot_chapter=False,
                  one_hot_utterance=False,
                  in_memory=True):
-        self.in_memory = in_memory
-        self.transforms = transforms
-        self.sr = sr
-        self.signal_length = signal_length
-        self.precision = precision
+        super(LibriSpeechH5py, self).__init__(transforms, sr, signal_length, precision, one_hot_all, in_memory)
 
         self.hpy_file = None
         f = h5py.File(dataset_path, 'r')
@@ -52,7 +35,6 @@ class LibriSpeechH5py(data.Dataset):
 
         self.n = self.speaker.shape[0]
 
-        self.one_hot_all = one_hot_all
         self.one_hot_speaker = one_hot_speaker
         self.one_hot_chapter = one_hot_chapter
         self.one_hot_utterance = one_hot_utterance
@@ -75,19 +57,6 @@ class LibriSpeechH5py(data.Dataset):
     def __exit__(self, exc_type, exc_value, traceback):
         if self.hpy_file is not None:
             self.hpy_file.close()
-
-    def __do_transform(self, sound):
-        if self.transforms:
-            trans_sig = self.transforms(sound.reshape((1, -1, 1)))
-            sound = tensor_to_numpy(trans_sig)
-
-        return sound
-
-    def __do_one_hot(self, id, encoder):
-        return encoder(np.array([id]).reshape((-1, 1))).toarray()[0, :]
-
-    def __len__(self):
-        return self.n
 
     def __getitem__(self, index):
         sound, speaker, chapter, utterance = self.sound[index], self.speaker[index], \
